add: 직업기초 능력평가(NCS) 영역 인증제 점수 추가 API 구현

+++ src/main/kotlin/com/team/incube/gsmc/v3/domain/score/calculator/impl/JlptScoreCalculator.kt
@@ -53,10 +53,10 @@ class JlptScoreCalculator : CategoryScoreCalculator() {
         return min(maxJlptScore + bonusScore, 10)
     }
 
-    private fun convertGradeToScore(grade: Int?): Int {
+    private fun convertGradeToScore(grade: Double?): Int {
         if (grade == null) return 0
 
-        return when (grade) {
+        return when (grade.toInt()) {
             1 -> 10
 
             // N1
